BufferGetLSNAtomic: fix Coverity complaint about negative array index
BufferGetLSNAtomic accepts local (negative) buffer values, and simply
exits early if it sees one. But it does index into the BufferDescriptors
array using this invalid negative value, which makes Coverity complain.

We're taking the address of the invalid location and not using it in the
local case, so this likely doesn't have any real ramifications. But I
can't find a clear statement that &array[invalid_index] is actually
well-defined across all platforms/compilers, and someone in the future
might miss that bufHdr isn't always pointing to something sane, so it
seems prudent to just fix this.