Convert GPDB-specific GUCs to the new "enum" type
Several GUCs are simply enumerated strings that are parsed into integer
types behind the scenes. As of 8.4, the GUC system recognizes a new
type, enum, which will do this for us. Move as many as we can to the new
system.

As part of this,
- gp_idf_deduplicate was changed from a char* string to an int, and new
  IDF_DEDUPLICATE_* macros were added for each option
- A new CODEGEN_OPTIMIZATION_LEVEL_UNSET option was added for the case
  where USE_CODEGEN is not defined. (XXX note that this allows a user to
  manually set codegen_optimization_level to the empty string, which
  will cause the system to fail...!)

A couple of GUCs that *could* be represented as enums
(optimizer_minidump, gp_workfile_compress_algorithm) have been
purposefully kept with the prior system because they require the GUC
variable to be something other than an integer anyway.

Signed-off-by: Jacob Champion <pchampion@pivotal.io>