CreateFakeRelcacheEntry: remove dead standby-filespace code
The directory creation code here seems to be duplicated in multiple
places. Since tests are passing now, we're not sure if it's necessary.
It seems to only apply to a standby master, when using non-default
filespaces, and we haven't found a good test case to reproduce the
initial issue that it was supposed to solve -- it seems like originally,
there was a problem where the filespace hashtable was not initialized on
the standby.

Since things seem to be working as intended now, get rid of the dead
code.

Signed-off-by: Jacob Champion <pchampion@pivotal.io>