WIP: Rebuild a failed primary into a new mirror.
Issue: the newly created mirror fails to start because of its pg_control state
is DB_IN_STANDBY_NEW_TLI_SET and that trips an assertion in StartupXLOG()
xlog.c:6567.  The logic to set the state to DB_IN_PRODUCTION was in pass4,
according to the comment at xlog.c:7440.  The promotion of the original mirror
didn't complete and it was left in state = DB_IN_STANDBY_NEW_TLI_SET.  During
basebackup, this state gets copied to the new mirror that was rebuilt.  And the
new mirror fails to start due to the assertion failure.

Next:
* Fix the promotion so that state is DB_IN_PRODUCTION at the end of promotion.
* New isolation2 test using the rebuild operation to restore cluster after promotion.

Author: Xin Zhang <xzhang@pivotal.io>
Author: Asim R P <apraveen@pivotal.io>