run_sql_command: validate unless the caller asks for the return code
We're seeing cases where a SQL command isn't even executing, but the
caller is continuing as if nothing is wrong. This makes debugging
difficult. Instead of hardcoding Command.run(validateAfter=False),
assume that we should only ignore the return code if the caller is
checking it with a results dict.

Author: Jacob Champion <pchampion@pivotal.io>
Author: Ashwin Agrawal <aagrawal@pivotal.io>