get_rel_infos: handle sequences during 8.2->8.3 upgrade
When upgrading to 8.4, upstream's pg_upgrade refuses to copy old
sequence data due to a change in columns. However, that check also
prevented copying of sequence data during an 8.2->8.3 upgrade --
Postgres doesn't care about this case, but GPDB does.

Don't omit the copy if we're upgrading to 8.3, and enable the 8.2 heap
upgrade for sequence tables so that the page makes sense in the new
database.

(cherry picked from commit 907803c02bf408c298ba10b187d7844c81255686)