pg_dump: work around hang during external partition exchange
The new optimized query introduced in commit 2ea6aec3 isn't as optimized
as we'd like if enable_nestloop is on -- released versions of GPDB craft
a plan that runs in polynomial time, and the dump grinds to a halt.

Disable enable_nestloop temporarily as a workaround. This code may need
to be somewhat rewritten to deal with external subpartition leaves
anyway, in which case we may be able to get rid of some of the
problematic JOINs.

Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>
(cherry picked from commit c4a7b25143d727d01a7545e6c4a29c5e6edbd90c)