startSegments: correctly track segment status
If gpsegstart returned a 0 or 1 exit code, but didn't provide a status
for every segment, the __processStartOrConvertCommands() helper would
happily ignore that, and we'd later assert when the number of successes
and failures didn't add up to the number of segments.

Keep a checklist of segments we've processed status for, and mark any
missing segments as having failed.