Stop connecting in utility mode on pg_dumpall
Starting with the 9.3 merge the arguments for the global connection to master
are being used to create the connections to all databases during
pg_dump.  This is causing pg_dump to not retrieve any data present in
the segments because we set utility mode on the global connection.

We reverted the utility mode change to align to upstream as we are not
required to be in utility mode to retrieve the information that we need
from master.