squash! Add pg_upgrade tests to master pipeline
Try to fix up the grouping dependencies so that they're logically
consistent. Require both ICW and CLI to be included to enable pg_upgrade
tests, and make sure that the sqldump resource is defined even if CLI
isn't specified.

This doesn't yet fix the case where a dev pipeline is generated with
only CLI, but it gets us a little closer.

Co-authored-by: Taylor Vesely <tvesely@pivotal.io>